DEPLOY: Prep extensions JsonConfig, Kartographer, Math, Score
This downloads the extension code and also prerequisites
but does not enable them yet (some db changes are needed).

This also creates a util-fether container that can be used for grabbing
and unzipping which saves ~1min of the code mediawiki image build
time...

This also removes the "evil" hack as it is no longer needed.

Touches #64 #62 #58